New ShoppingCart crud (Deleting problems)

+++ src/main/java/com/uq/jokievents/service/ShoppingCartService.java
@@ -1,8 +1,11 @@
 package com.uq.jokievents.service;
 
+import com.uq.jokievents.model.Client;
 import com.uq.jokievents.model.ShoppingCart;
 import com.uq.jokievents.repository.ShoppingCartRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,8 +22,13 @@ public class ShoppingCartService {
      *
      * @return a list of all ShoppingCarts objects in the db
      */
-    public List<ShoppingCart> findAll() {
-        return shoppingCartRepository.findAll();
+    public ResponseEntity<?> findAll() {
+        try {
+            List<ShoppingCart> shoppingCarts = shoppingCartRepository.findAll();
+            return new ResponseEntity<>(shoppingCarts, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed ShoppingCarts request", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -29,8 +37,17 @@ public class ShoppingCartService {
      * @param id unique identifier of the ShoppinCart object
      * @return an Optional containing the ShoppingCart if found, empy Optional if not
      */
-    public Optional<ShoppingCart> findById(String id) {
-        return shoppingCartRepository.findById(id);
+    public ResponseEntity<?> findById(String id) {
+        try {
+            Optional<ShoppingCart> shoppingCart = shoppingCartRepository.findById(id);
+            if (shoppingCart.isPresent()) {
+                return new ResponseEntity<>(shoppingCart.get(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("ShoppingCart not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed ShoppingCart request", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -39,8 +56,35 @@ public class ShoppingCartService {
      * @param shoppingCart object to be saved or updated
      * @return the saved or updated ShoppingCart object
      */
-    public ShoppingCart save(ShoppingCart shoppingCart) {
-        return shoppingCartRepository.save(shoppingCart);
+    public ResponseEntity<?> create(ShoppingCart shoppingCart) {
+        try {
+            ShoppingCart createdShoppingCart = shoppingCartRepository.save(shoppingCart);
+            return new ResponseEntity<>(createdShoppingCart, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to create ShoppingCart", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    /**
+     * Update an existing client by id
+     *
+     * @param id the identifierof the client to be update
+     * @param shoppingCart the shoppingCart client object
+     * @return a ResponseEntity containing the updated shoppingCart objec and an HTTP status
+     */
+    public ResponseEntity<?> update(String id, ShoppingCart shoppingCart) {
+        try {
+            Optional<ShoppingCart> existingShoppingCart = shoppingCartRepository.findById(id);
+            if (existingShoppingCart.isPresent()) {
+                shoppingCart.setId(id);
+                ShoppingCart updatedShoppingCart = shoppingCartRepository.save(shoppingCart);
+                return new ResponseEntity<>(updatedShoppingCart, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("ShoppingCart not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to update shoppingCart", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -48,7 +92,17 @@ public class ShoppingCartService {
      *
      * @param id the unique identifier of the object
      */
-    public void deleteById(String id) {
-        shoppingCartRepository.deleteById(id);
+    public ResponseEntity<?> deleteById(String id) {
+        try {
+            Optional<ShoppingCart> existingShoppingCart = shoppingCartRepository.findById(id);
+            if (existingShoppingCart.isPresent()) {
+                shoppingCartRepository.deleteById(id);
+                return new ResponseEntity<>("ShoppingCart deleted", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("ShoppingCart not found", HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>("Failed to delete ShoppingCart", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
